requirements.txt
Flask==2.0.1
pandas==1.3.0
fpdf2==2.5.4
werkzeug==2.0.1